refact : header folder renamed to Header, redundant style, type files removed from Header folder
