+++ src/App.tsx
@@ -1,5 +1,5 @@
 import { Outlet } from "react-router-dom";
-import Header from "./components/header";
+import Header from "./components/Header";
 
 function App() {
     return (
